+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # hastons
 
+## 1.1.2
+
+### Patch Changes
+
+- fix(analytics): try using fetch
+
+## 1.1.1
+
+### Patch Changes
+
+- fix(analytics): add CORS headers to track API and fix pathname cleaning
+
 ## 1.1.0
 
 ### Minor Changes